+++ src/gamspy/_model_instance.py
@@ -170,7 +170,7 @@ class ModelInstance:
     def solve(
         self,
         solver: str | None,
-        given_options: ModelInstanceOptions | dict | None = None,
+        given_options: ModelInstanceOptions | None = None,
         output: io.TextIOWrapper | None = None,
     ):
         # get options from dict
@@ -279,7 +279,7 @@ class ModelInstance:
     def _prepare_options(
         self,
         solver: str | None,
-        given_options: ModelInstanceOptions | dict | None,
+        given_options: ModelInstanceOptions | None,
     ) -> tuple[GamsModelInstanceOpt | None, SymbolUpdateType]:
         update_type = SymbolUpdateType.BaseCase
         options = GamsModelInstanceOpt()