+++ CHANGELOG.md
@@ -10,6 +10,8 @@ GAMSPy 1.0.1
   - Add test for the frozen solve with non-scalar symbols.
   - Add a test to verify the behaviour of equation redefinition with definition argument.
   - Test the usage of a license that is in one of the default paths.
+  - Fix the issue related to reading equation records from a gdx file.
+  - Add tests to verify the records after reading them from a gdx file.
 
 GAMSPy 1.0.0
 ------------